requirements.txt
azure-core==1.21.1
azure-storage-blob==12.9.0
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
idna==3.3
isodate==0.6.1
jmespath==0.10.0
msrest==0.6.21
oauthlib==3.1.1
pycparser==2.21
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
